implement permissions

+++ formula1-tournament-api/Controllers/RaceController.cs
@@ -10,17 +10,19 @@ namespace formula1_tournament_api.Controllers
     public class RaceController : Controller
     {
         private IRace _raceService;
+        private IUserSeason _userSeasonService;
 
-        public RaceController(IRace raceService)
+        public RaceController(IRace raceService, IUserSeason userSeasonService)
         {
             _raceService = raceService;
+            _userSeasonService = userSeasonService;
         }
 
-        [HttpGet]
+        [HttpGet("{seasonId}")]
         [EnableQuery]
-        public async Task<IActionResult> Get()
+        public async Task<IActionResult> Get(Guid seasonId)
         {
-            var result = await _raceService.GetAllRaces();
+            var result = await _raceService.GetAllRacesBySeasonId(seasonId);
             if (result.IsSuccess)
             {
                 return Ok(result.Races);
@@ -28,8 +30,8 @@ namespace formula1_tournament_api.Controllers
             return NotFound(result.ErrorMessage);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> Get(Guid id)
+        [HttpGet("{seasonId}/{id}")]
+        public async Task<IActionResult> Get(Guid seasonId, Guid id)
         {
             var result = await _raceService.GetRaceById(id);
             if (result.IsSuccess)
@@ -39,9 +41,11 @@ namespace formula1_tournament_api.Controllers
             return NotFound(result.ErrorMessage);
         }
 
-        [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Race race)
+        [HttpPost("{seasonId}")]
+        public async Task<IActionResult> Post(Guid seasonId, [FromBody] Race race)
         {
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
             var result = await _raceService.AddRace(race);
             if (result.IsSuccess)
             {
@@ -50,9 +54,11 @@ namespace formula1_tournament_api.Controllers
             return BadRequest(result.ErrorMessage);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Put(Guid id, [FromBody] Race race)
+        [HttpPut("{seasonId}/{id}")]
+        public async Task<IActionResult> Put(Guid seasonId, Guid id, [FromBody] Race race)
         {
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
             var result = await _raceService.UpdateRace(id, race);
             if (result.IsSuccess)
             {
@@ -61,9 +67,11 @@ namespace formula1_tournament_api.Controllers
             return BadRequest(result.ErrorMessage);
         }
 
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(Guid id)
+        [HttpDelete("{seasonId}/{id}")]
+        public async Task<IActionResult> Delete(Guid seasonId, Guid id)
         {
+            if (!_userSeasonService.HasPermission(new Guid(User.Identity.Name), seasonId))
+                return StatusCode(StatusCodes.Status403Forbidden);
             var result = await _raceService.DeleteRace(id);
             if (result.IsSuccess)
             {
